Add files via upload
Finished penalty calculator. Method takes the finalSolution array and parses through it to grab the element as well as index, then it uses both values to get the penalty value from the machine-penalties array. The value is then added onto the penaltiesValue variable and returned once the finalSolution array has been traversed.